refaktor middleware autht, authr
