+++ server/middlewares/errorHandlers.js
@@ -1,0 +1,5 @@
+module.exports = (err, req, res, next) => {
+    console.log (err)
+
+
+}
